// frontend/src/app/v2/core/enums/InputTypes.ts
export type InputType =
    | "label"
    | "text"
    | "url"
    | "number"
    | "money"
    | "date"
    | "boolean"
    | "image"
    | "password"
    | "enumerator"
    | "unit"
    | "type"
    | "group"
    | "form"
    | "enumerator_value"
    | "unit_value";

export enum InputTypes {
    'label' = 1,
    'text' = 2,
    'url' = 3,
    'number' = 4,
    'money' = 5,
    'date' = 6,
    'boolean' = 7,
    'image' = 8,
    'password' = 9,
    'enumerator' = 10,
    'unit' = 11,
    'type' = 12,
    'group' = 13,
    'form' = 14,
    'enumerator_value' = 15,
    'unit_value' = 16
}